Version 2
I added a powerful functions to search in bibliographic files (*.bib) for entries which match a vim regexp. One choose which entries are shown (articles,books, etc..., but also author,title,etc...). The function can easily copy the bibfield key to a register or directly to the tex source file.

I also supply a syntax file for output of this search function which is a small modification of the standard syntax file for bib files.